Map invalid register to the physical null address
This solves problems with Immediate type instructions, which do not
set a register address for e.g. Rb, which causes INVALID_REG to be
used, which was set to register 1F. This is correct for Alpha, but
for MIPS it triggers an illegal instruction error.

Define the invlid register index to 0 on MIPS targets, which is
cleaner than setting them explicitly in the decode logic or
translating register 31 to zero in GetRegisterClass.